examples/scripts/key-generation/requirements.txt
python-gnupg>=0.5.0
yubikey-manager>=5.0.0
cryptography>=41.0.0
typing-extensions>=4.5.0